서비스 인터페이스 추가
